% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Procedures.R
\name{compile_sbtests}
\alias{compile_sbtests}
\title{Compile the output of Sup Wald test}
\usage{
compile_sbtests(x, digits = 3)
}
\arguments{
\item{x}{\code{sbtests} class object}

\item{digits}{number of decimal places displayed}
}
\value{
class \code{sbtests} x with appended 2 data frames:
\itemize{
\item{supF1 A data frame contains SupF tests statistics of \code{0} versus \code{m} breaks, where
\code{m} is maximum number of breaks considered in \code{x} with critical values at
\verb{\{10\%,5\%,2.5\%,1\%\}} level}
\item{UDMax A data frame contains Double Max test statistics with with critical values at
\verb{\{10\%,5\%,2.5\%,1\%\}} level}
}
}
\description{
\code{compile_sbtests} formats the output of the \code{sbtests} into 2 tables
\itemize{
\item supF1 table containing Sup F tests of \code{0} versus \code{1} to \code{m} breaks
with critical values of the corresponding tests
\item}
}
