% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Procedures.R
\name{compile_seqtests}
\alias{compile_seqtests}
\title{Compile the output of sequential Sup Wald test}
\usage{
compile_seqtests(x)
}
\arguments{
\item{x}{\code{seqtests} class object}
}
\value{
class \code{seqtests} list \code{x} with appended data frame \code{sfl} containing the
sequential SupF test statistics with critical values at
\verb{\{10\%,5\%,2.5\%,1\%\}} level
}
\description{
\code{compile_seqtests} formats the output of the \code{seqtests} class object to 1 table
\itemize{
\item sfl table containing sequential sup F tests statistics
of \code{l} versus \code{l+1} for l in \code{1} to \code{m} breaks
with critical values of the corresponding tests at
\verb{\{1\%, 2.5\%, 5\%, 10\%\}} significance levels
}
}
