% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Variance.R
\name{correct}
\alias{correct}
\title{Heteroskedasticy and autocorrelation consistency correction for residuals}
\usage{
correct(reg, res, prewhit)
}
\arguments{
\item{reg}{matrix of regressors}

\item{res}{matrix of estimated residuals}

\item{prewhit}{Option of using prewhitening process. If \code{1}, an AR(1)
process will be used to filter. If \code{0}, skipped the filtering process}
}
\value{
hac Heteroskedasticy and autocorrelation consistent errors
}
\description{
`hac()` corrects the estimated errors based on options of prewhitening
using a AR(1) process estimation of error terms to obtain
heteroskedasticy and autocorrelation consistency (HAC) errors with automatic
bandwith and kernel similar to Andrews, 1994
}
