% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structbreaks.R
\name{correct1}
\alias{correct1}
\title{Compute robust standard errors in joint test of sc (i.e., Perron et al. 2021)}
\usage{
correct1(vmat, vmata, prewhit, typekb)
}
\description{
Compute robust standard errors in joint test of sc (i.e., Perron et al. 2021)
}
\references{
Perron, Pierre, Yohei Yamamoto, and Jing Zhou (2020), "Testing Jointly for Structural Changes in the Error Variance and Coefficients of a Linear Regression Model" \emph{Quantitative Economics}, vol 11, 1019-1057.
}
