% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.R, R/data.R
\name{dating}
\alias{dating}
\title{Calculate optimal break points for pure structural change model}
\usage{
dating(y, z, h, m, q, bigT)

dating(y, z, h, m, q, bigT)
}
\arguments{
\item{y}{dependent variable}

\item{z}{matrix of independent variables}

\item{h}{minimum length of segment}

\item{m}{maximum number of breaks}

\item{q}{number of regressors z}

\item{bigT}{sample period T}
}
\value{
A list containing the following components:
\item{glb}{minimum global SSR}
\item{datevec}{Vector of dates (optimal minimizers)}
\item{bigvec}{Associated SSRs}

A list containing the following components:
\item{glb}{minimum global SSR}
\item{datevec}{Vector of dates (optimal minimizers)}
\item{bigvec}{Associated SSRs}
}
\description{
Function computes break points that globally minimizes SSR via dynamic programming approach.
To avoid recursion depth increases as number of breaks in the model increases, a temporary
array is used to store optimal partition with corresponding SSR for all permissible
subsamples for all 1:m-1 breaks. For m breaks, the problem became finding where to insert
the last feasible segments into a subsample to obtain minimum SSR for 1 to T

Function computes break points that globally minimizes SSR
}
