% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dating_MLE}
\alias{dating_MLE}
\title{Compute global break dates using log-likelihood}
\usage{
dating_MLE(y, z, q, x, p, h, m, bigt)
}
\description{
This is the main procedure which calculates the break points that globally maximize the loglikelihood function. It returns optimal dates and associated log likelihood for all numbers of breaks less than or equal to m.
}
\references{
Perron, Pierre, Yohei Yamamoto, and Jing Zhou (2020), "Testing Jointly for Structural Changes in the Error Variance and Coefficients of a Linear Regression Model" \emph{Quantitative Economics}, vol 11, 1019-1057.
}
