% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{determineBreaks}
\alias{determineBreaks}
\title{Determine number of breaks & dates}
\usage{
determineBreaks(y, M, N, z, x, con)
}
\description{
This function determines the number of break and their dates using sequential and UDmax procedures.
}
\references{
Perron, Pierre, Yohei Yamamoto, and Jing Zhou (2020), "Testing Jointly for Structural Changes in the Error Variance and Coefficients of a Linear Regression Model" \emph{Quantitative Economics}, vol 11, 1019-1057.
}
