% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Procedures.R
\name{doglob}
\alias{doglob}
\title{Global SSR minimizers procedure}
\usage{
doglob(y, z, x, m, eps, eps1, maxi, fixb, betaini, printd)
}
\arguments{
\item{eps}{convergence criterion for iterative recursive computation}

\item{eps1}{trimming level}

\item{maxi}{maximum number of iterations}

\item{fixb}{option to use fixed initial input \eqn{\beta}. If \code{1},
the model will use values given in \code{betaini}. If \code{0}, betaini is skipped}

\item{betaini}{Initial \eqn{beta_0} to use in estimation}

\item{printd}{option to print results of iterations for partial change model}

\item{y_name}{dependent variables in matrix form}

\item{z_name}{matrix of independent variables with coefficients are allowed to change across
regimes}

\item{x_name}{matrix of independent variables with coefficients constant across regimes}
}
\value{
A list containing the following components:
\itemize{
\item{glb} {Minimum global SSR}
\item{datevec} {Vector of dates (optimal minimizers)}
\item{bigvec} {Associated SSRs with possible break dates combination}}
}
\description{
A helper function to identify if the estimated break model is i) pure change or ii)
partial change model. The procedure then calls appropriate functions \link{} to estimate
the pure change model and \link{} to estimate the partial change model. This helper function
is required for supF, UDMax, WDMax and supF(l+1|l) test functions invoked via \link{}
}
