% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Procedures.R
\name{doseqtests}
\alias{doseqtests}
\title{Sequential Sup F tests}
\usage{
doseqtests(
  y_name,
  z_name = NULL,
  x_name = NULL,
  data,
  m = 5,
  eps = 1e-05,
  eps1 = 0.15,
  maxi = 10,
  fixb = 0,
  betaini = 0,
  printd = 0,
  prewhit = 1,
  robust = 1,
  hetdat = 1,
  hetvar = 1,
  hetq = 1,
  hetomega = 1,
  const = 1
)
}
\arguments{
\item{y_name}{name of dependent variable in the data set}

\item{z_name}{name of independent variables in the data set which coefficients are allowed to change
across regimes. \code{default} is a vector of 1 (Mean-shift model)}

\item{x_name}{name of independent variables in the data set which coefficients are constant across
regimes. \code{default} is `NULL`}

\item{data}{name of data set used}

\item{m}{maximum number of breaks}

\item{eps}{convergence criterion for recursive calculations (For partial change model ONLY)}

\item{eps1}{value of trimming (in percentage) for the construction
and critical values. Minimal segment length `h` will be set
at \code{default} = int(\code{eps1}*T) (T is total sample size).
 There are five options:
\itemize{
\item `eps1=0.05` Maximal value of `m` = 10.
\item `eps1=0.10` Maximal value of `m` = 8.
\item `eps1=.15` Maximal value of `m` = 5.
\item `eps1=.20` Maximal value of `m` = 3.
\item `eps1=.25` Maximal value of `m` = 2.
\item `eps1=0` is not allowed. The test is undefined for no trimming level.
}}

\item{maxi}{number of maximum iterations for recursive calculations of finding
global minimizers.\code{default} = 10 (For partial change model ONLY)}

\item{fixb}{option to use fixed initial input \eqn{\beta}. If \code{1},
the model will use values given in \code{betaini}. If \code{0}, betaini is skipped}

\item{betaini}{Initial \eqn{beta_0} to use in estimation (Must be a `p x 1` matrix, where p is number of x variables)}

\item{printd}{Print option for model estimation. \code{default} = 0, to
suppress intermediate outputs printing to console}

\item{prewhit}{set to \code{1} to apply AR(1) prewhitening prior to estimating
the long run covariance matrix.}

\item{robust}{set to \code{1} to allow for heterogeneity
and autocorrelation in the residuals, \code{0} otherwise.
The method used is \emph{Andrews(1991)} automatic bandwidth with AR(1) approximation with quadratic
kernel. Note: Do not set to \code{1} if lagged dependent variables are
included as regressors.}

\item{hetdat}{option for the construction of the F tests. Set to 1 if want to
allow different moment matrices of the regressors across segments.
If \code{hetdat} = \code{0}, the same moment matrices are assumed for each segment
and estimated from the ful sample. It is recommended to set
\code{hetdat}=\code{1} if number of regressors \code{x} > \code{0}.}

\item{hetvar}{option for the construction of the F tests.Set to \code{1}
if users want to allow for the variance of the residuals to be different across segments.
If \code{hetvar}=\code{0}, the variance of the residuals is assumed constant
across segments and constructed from the full sample. \code{hetvar}=\code{1} when \code{robust} =\code{1})}

\item{hetq}{used in the construction of the confidence intervals for the break
dates. If \code{hetq}=\code{0}, the moment matrix of the data is assumed identical
across segments}

\item{hetomega}{used in the construction of the confidence intervals for the break
dates. If \code{hetomega}=\code{0}, the long run covariance matrix of zu is
assumed identical across segments
(the variance of the errors u if \code{robust}=\code{0})}

\item{const}{indicates whether the regression model include an
intercept changing across regimes. Default value is 1}
}
\value{
A list that contains following:
\describe{
\item{supfl}{SupF(l+1|l) test statistics.}
\item{cv}{Critical values for SupF(l+1|l) test.}}
}
\description{
`doseqtests()` computes the sequential sup F tests of l versus l+1 for l from
1 to m with each corresponding null hypothesis of maximum number of break is l
and alternative hypothesis is l+1. The l breaks under the null hypothesis
are taken from the global minimization estimation
}
\examples{
doseqtests('inf',c('inflag','lbs','inffut'),data=nkpc,prewhit=0)

}
