% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structbreaks.R
\name{estimbr}
\alias{estimbr}
\title{Estimate FGLS}
\usage{
estimbr(y, z, q, x, p, bigt, K, brk, R, n, brv, rest)
}
\description{
procedure to estimate coefficients and residuals by FGLS given break dates in coefficients and variance.
}
\references{
Perron, Pierre, Yohei Yamamoto, and Jing Zhou (2020), "Testing Jointly for Structural Changes in the Error Variance and Coefficients of a Linear Regression Model" \emph{Quantitative Economics}, vol 11, 1019-1057.
}
