% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{estimdl}
\alias{estimdl}
\title{Estimate model}
\usage{
estimdl(y, m, n, z, x = matrix(0, 0, 0), control = list())
}
\arguments{
\item{y}{A (\code{T x 1}) vector with endogenous variable.}

\item{m}{An integer determining the number of breaks in coefficients.}

\item{n}{An integer determining the number of breaks in variance.}

\item{z}{A (\code{T x q}) matrix with explanatory variables subject to change.}

\item{control}{List with options which include
\itemize{
  \item{\code{robust}: }{set to 1 if want to allow for heterogeneity and autocorrelation in the residuals, 0 otherwise. The method used is Andrews(1991) automatic bandwidth with AR(1) approximation and the quadratic kernel. Note: Do not set to 1 if lagged dependent variables are included as regressors. Default is \code{TRUE}.}
  \item{\code{prewhit}: }{set to 1 if want to apply AR(1) prewhitening prior to estimating the long run covariance matrix. Default is \code{TRUE}.}
  \item{\code{hetdat}: }{option for the construction of the F tests. Set to 1 if want to allow different moment matrices of the regressors across segments. If hetdat=0, the same moment matrices are assumed for each segment and estimated from the ful sample. It is recommended to set hetdat=1 if p>0. Default is \code{TRUE}.}
  \item{\code{hetvar}: }{option for the construction of the F tests.Set to 1 if want to allow for the variance of the residuals to be different across segments. If hetvar=0, the variance of the residuals is assumed constant across segments and constructed from the ful sample. This option is not available when robust=1. Default is \code{TRUE}.}
  \item{\code{hetomega}: }{used in the construction of the confidence intervals for the break dates. If hetomega=0, the long run covariance matrix of zu is assumed identical across segments (the variance of the errors u if robust=0). Default is \code{TRUE}.}
  \item{\code{hetq}: }{used in the construction of the confidence intervals for the break dates. If hetq=0, the moment matrix of the data is assumed identical across segments. Default is \code{TRUE}.}
  \item{\code{doglobal}: }{set to 1 if want to cal the procedure to obtain global minimizers. Default is \code{TRUE}.}
  \item{\code{hetomega}: }{used in the construction of the confidence intervals for the break dates. If hetomega=0, the long run covariance matrix of zu is assumed identical across segments (the variance of the errors u if robust=0Default is \code{TRUE}.}
  \item{\code{hetq}: }{used in the construction of the confidence intervals for the break dates. If hetq=0, the moment matrix of the data is assumed identical across segments. Default is \code{TRUE}.}
  \item{\code{}: }{Default is \code{TRUE}.}
}}
}
\description{
Model estimation with known or unknown breaks
}
\details{
Note: This code is a translation of the one originally written by Pierre 
Perron and Yohei Yamamoto for MATLAB. Original code files can be found on 
Pierre Perron's website: https://blogs.bu.edu/perron/codes/
}
\references{
Perron, Pierre, Yohei Yamamoto, and Jing Zhou (2020), "Testing Jointly for Structural Changes in the Error Variance and Coefficients of a Linear Regression Model" \emph{Quantitative Economics}, vol 11, 1019-1057.
}
