% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_methods.R
\name{getcv}
\alias{getcv}
\title{Get critical values using equations from Bai Perron (2003)}
\usage{
getcv(alpha, q, k, eps = 0.1)
}
\description{
Get critical values using equations from Bai Perron (2003)
}
