% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structbreaks.R
\name{getdmax4}
\alias{getdmax4}
\title{Get critical values for UDmaxLRT_4}
\usage{
getdmax4(alpha, q, trm, M, N)
}
\description{
Get critical values for UDmaxLRT_4
}
\references{
Perron, Pierre, Yohei Yamamoto, and Jing Zhou (2020), "Testing Jointly for Structural Changes in the Error Variance and Coefficients of a Linear Regression Model" \emph{Quantitative Economics}, vol 11, 1019-1057.
}
