% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inflation}
\alias{inflation}
\title{Inflation series (quarterly) from 1959Q1 - 2018Q4}
\format{
\describe{Annual change of the quarterly US inflation rate from 1959Q1 – 2018Q4 used in Perron, Yamamoto, and Zhou (2020)
  \item{inflation}{US inflation series}
}
}
\source{
\url{https://blogs.bu.edu/perron/codes/}
}
\usage{
inflation
}
\description{
Inflation series (quarterly) from 1959Q1 - 2018Q4
}
\references{
Perron, Pierre, Yohei Yamamoto, and Jing Zhou (2020), "Testing Jointly for Structural Changes in the Error Variance and Coefficients of a Linear Regression Model" \emph{Quantitative Economics}, vol 11, 1019-1057.
}
\keyword{datasets}
