% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structbreaks.R
\name{interval}
\alias{interval}
\title{Confidence intervals for breaks}
\usage{
interval(y, z, zbar, b, q, m, robust, prewhit, hetomega, hetq, x, p)
}
\description{
This procedure that computes confidence intervals 
for the break dates based on the "shrinking shifts" asymptotic framework.
}
\details{
Note: This code is an adaptation of the one originally written by Yohei 
Yamamoto and Pierre Perron for MATLAB. Original codes can be found on 
Pierre Perron's website: https://blogs.bu.edu/perron/codes/
}
\references{
Bai, Jushan & Pierre Perron (1998), "Estimating and Testing Linear Models with Multiple Structural Changes," \emph{Econometrica}, vol 66, 47-78.

Bai, Jushan & Pierre Perron (2003), "Computation and Analysis of Multiple Structural Change Models," \emph{Journal of Applied Econometrics}, 18, 1-22.
}
