% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_methods.R
\name{invpd}
\alias{invpd}
\title{Inverse of positive definite matrix}
\usage{
invpd(x)
}
\description{
Inverse of positive definite matrix
}
\references{
Perron, Pierre, Yohei Yamamoto, and Jing Zhou (2020), "Testing Jointly for Structural Changes in the Error Variance and Coefficients of a Linear Regression Model" \emph{Quantitative Economics}, vol 11, 1019-1057.
}
