% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structbreaks.R
\name{isposdef}
\alias{isposdef}
\title{Determine if square matrix is positive definite}
\usage{
isposdef(A)
}
\description{
Determine if square matrix is positive definite
}
\references{
Perron, Pierre, Yohei Yamamoto, and Jing Zhou (2020), "Testing Jointly for Structural Changes in the Error Variance and Coefficients of a Linear Regression Model" \emph{Quantitative Economics}, vol 11, 1019-1057.
}
