% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Variance.R
\name{jhatpr}
\alias{jhatpr}
\title{Long-run covariance matrix computation}
\usage{
jhatpr(vmat)
}
\arguments{
\item{vmat}{variance matrix}
}
\value{
jhat Long run covariance matrix
}
\description{
The procedure to compute the long run covariance matrix \code{jhat} based
on variance matrix \code{vmat}
}
