% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structbreaks.R
\name{jhatpr1}
\alias{jhatpr1}
\title{Compute robust standard errors for correct1() function}
\usage{
jhatpr1(vmat, vmata)
}
\description{
Compute robust standard errors for correct1() function
}
\references{
Perron, Pierre, Yohei Yamamoto, and Jing Zhou (2020), "Testing Jointly for Structural Changes in the Error Variance and Coefficients of a Linear Regression Model" \emph{Quantitative Economics}, vol 11, 1019-1057.
}
