\name{mbreaks-package}
\alias{mbreaks-package}
\alias{mbreaks}
\docType{package}
\title{
\packageTitle{mbreaks}
}
\description{
\packageDescription{mbreaks}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mbreaks}
\packageIndices{mbreaks}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{mbreaks}

Maintainer: \packageMaintainer{mbreaks}
}
\references{
Bai, Jushan & Pierre Perron (1998), Estimating and Testing Linear Models with Multiple Structural Changes, Econometrica, vol 66, 47-78. https://doi.org/10.2307/2998540

Bai, J. and Perron, P. (2003), Computation and analysis of multiple structural change models, Journal of Applied Econometrics, 18: 1-22. https://doi.org/10.1002/jae.659

Bai, J. and Perron, P. (2003), Critical values for multiple structural change tests, The Econometrics Journal, 6: 72-78. https://doi.org/10.1111/1368-423X.00102

Perron, Pierre, Yohei Yamamoto, and Jing Zhou (2020), Testing Jointly for Structural Changes in the Error Variance and Coefficients of a Linear Regression Model, Quantitative Economics, vol 11, 1019-105. https://doi.org/10.3982/QE1332

}
