% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mlef}
\alias{mlef}
\title{MLE function}
\usage{
mlef(start, y, z, q, x, p, h, last)
}
\description{
MLE function
}
\references{
Perron, Pierre, Yohei Yamamoto, and Jing Zhou (2020), "Testing Jointly for Structural Changes in the Error Variance and Coefficients of a Linear Regression Model" \emph{Quantitative Economics}, vol 11, 1019-1057.
}
