% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nkpc}
\alias{nkpc}
\title{New Keynesian Phillips curve data}
\format{
\subsection{\code{nkpc}}{

A data frame with 7,240 rows and 60 columns:
\describe{
\item{inf}{Inflation rate}
\item{inffut}{Expected inflation rate, take value of inflation rate of next period}
\item{year}{Year}
...
}
}
}
\source{
\url{https://www.who.int/teams/global-tuberculosis-programme/data}
}
\usage{
nkpc
}
\description{
A subset of data from the World Health Organization Global Tuberculosis
Report ...
}
\keyword{datasets}
