% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structbreaks.R
\name{numcase2}
\alias{numcase2}
\title{Determine number of cases to check given number of breaks in mean or var}
\usage{
numcase2(m, n)
}
\description{
replaces brcvcase() and numcase() in original MATLAB code.
}
\references{
Perron, Pierre, Yohei Yamamoto, and Jing Zhou (2020), "Testing Jointly for Structural Changes in the Error Variance and Coefficients of a Linear Regression Model" \emph{Quantitative Economics}, vol 11, 1019-1057.
}
