% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structbreaks.R
\name{OLS}
\alias{OLS}
\title{Estimate OLS with White robust standard errors}
\usage{
OLS(Y, X, add_constant)
}
\description{
This function estimates an ordinary least squares model with White robust 
standard error.
}
\details{
Note: This code is an adaptation of the one originally written by Yohei 
Yamamoto and Pierre Perron for MATLAB. Original codes can be found on 
Pierre Perron's website: https://blogs.bu.edu/perron/codes/
}
\references{
Bai, Jushan & Pierre Perron (1998), "Estimating and Testing Linear Models with Multiple Structural Changes," \emph{Econometrica}, vol 66, 47-78.

Bai, Jushan & Pierre Perron (2003), "Computation and Analysis of Multiple Structural Change Models," \emph{Journal of Applied Econometrics}, 18, 1-22.
}
