% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstDate.R
\name{onebp}
\alias{onebp}
\title{Optimal one break partition in partial structural change model}
\usage{
onebp(y, z, x, h, start, last)
}
\arguments{
\item{y}{dependent variables}

\item{z}{independent variables with coefficients
allowed to change across regimes}

\item{x}{independent variables with constant coefficients across regimes}

\item{h}{minimal length of segment}

\item{start}{initial date to search}

\item{last}{last date to search}
}
\value{
A list containing the following components
\item{ssrmin}{associated SSR of optimal break}
\item{dx}{optimal date (global minimizer)}
}
\description{
Function computes the optimal one break partition in partial structural
change model by searching over all possible breaks
}
