% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_methods.R
\name{parti}
\alias{parti}
\title{optimal break partitions for a given segment}
\usage{
parti(start, b1, b2, last, date_mat)
}
\description{
procedure to obtain an optimal one break partitions for a segment that
starts at date start and ends at date last. It returns the optimal break and the
associated SSR.
}
\details{
Note: This code is an adaptation of the one originally written by Yohei
Yamamoto and Pierre Perron for MATLAB. Original codes can be found on
Pierre Perron's website: https://blogs.bu.edu/perron/codes/
}
\references{
Bai, Jushan & Pierre Perron (1998), "Estimating and Testing Linear Models with Multiple Structural Changes," \emph{Econometrica}, vol 66, 47-78.

Bai, Jushan & Pierre Perron (2003), "Computation and Analysis of Multiple Structural Change Models," \emph{Journal of Applied Econometrics}, 18, 1-22.
}
