% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structbreaks.R
\name{parti_loglik}
\alias{parti_loglik}
\title{optimal break partitions for a given segment using log-likelihood}
\usage{
parti_loglik(start, b1, b2, last, bigvec, bigt)
}
\description{
procedure to obtain an optimal one break partitions for a segment that 
starts at date start and ends at date last. It returns the optimal break and the 
associated log-likelihood (parti2() in original MATLAB code)
}
\references{
Perron, Pierre, Yohei Yamamoto, and Jing Zhou (2020), "Testing Jointly for Structural Changes in the Error Variance and Coefficients of a Linear Regression Model" \emph{Quantitative Economics}, vol 11, 1019-1057.
}
