% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstDate.R
\name{partione}
\alias{partione}
\title{Optimal one break partitions with sequential method}
\usage{
partione(b1, b2, last, vssr, vssrev)
}
\arguments{
\item{last}{end of segment considered}

\item{vssr}{minimum SSRs of associated break date}

\item{vssrev}{recursive SSRs of the model}

\item{b_start}{first possible break date}

\item{b_end}{last possible break date}
}
\value{
A list containing the following components:
\item{ssrmin}{associated SSR of optimal break}
\item{dx}{optimal date (global minimizer)}
}
\description{
Function calculates an optimal one break partitions for a segment that
starts at date start and ends at date last. It returns the optimal break
and the associated SSR. Procedure used with the sequential method when
the T*(T+1)/2 vector of SSR is not computed.
}
\references{

}
