% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test.R
\name{pftest}
\alias{pftest}
\title{SupF test for 0 vs i breaks}
\usage{
pftest(y, z, i, q, bigT, datevec, prewhit, robust, x, p, hetdat, hetvar)
}
\arguments{
\item{y}{dependent variables}

\item{z}{independent variables with coefficients are allowed to change across
regimes}

\item{i}{number of breaks in the model}

\item{q}{number of \code{z} regressors}

\item{bigT}{sample period T}

\item{datevec}{\code{i} estimated dates from the model}

\item{prewhit}{Options for prewhitening process}

\item{robust, hetdat, hetvar}{options for assumptions on error terms}

\item{x}{independent variables with constant coefficients across regimes}

\item{p}{number of \code{x} regressors}
}
\value{
ftest SupF test results
}
\description{
Function compute the supF test statistics of testing procedure with
null hypothesis: no break versus alternative hypothesis: \code{i} breaks.
}
