% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structbreaks.R
\name{ploglik}
\alias{ploglik}
\title{Log likelihood given residuakls}
\usage{
ploglik(res, n, brv)
}
\description{
Log likelihood given residuakls
}
\references{
Perron, Pierre, Yohei Yamamoto, and Jing Zhou (2020), "Testing Jointly for Structural Changes in the Error Variance and Coefficients of a Linear Regression Model" \emph{Quantitative Economics}, vol 11, 1019-1057.
}
