% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl.R
\name{print.mdl}
\alias{print.mdl}
\title{Print information of \code{mbreaks} comprehensive procedure}
\usage{
\method{print}{mdl}(x, ...)
}
\arguments{
\item{x}{class \code{mdl} object}

\item{...}{further arguments passed to or from other methods}
}
\value{
No return value, only for printing \code{model}, \code{sbtests} and \code{seqtests} class objects
invoked during \code{mdl()}.
}
\description{
\code{print} prints the class \code{mdl} object with default showing only
certain procedures called
by \code{mdl()} function including: \code{seqtests} class object, \code{sbtests} class
object, and \code{model} class object using KT information criterion
}
\examples{
rate = mdl('rate',data=real)
print(rate)
}
