% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Procedures.R
\name{print.sbtests}
\alias{print.sbtests}
\title{Print Sup F and UDMax tests}
\usage{
\method{print}{sbtests}(x, ...)
}
\arguments{
\item{x}{class \code{sbtests} object}

\item{...}{further arguments passed to or from other methods}
}
\value{
No return value, only for printing formatted \code{sbtests} class object to console
}
\description{
\code{print} prints the following information from a \code{sbtests} class object:
\itemize{
\item table \code{supF1} reporting sup F tests of 0 versus \code{1} upto \code{m} breaks with critical values for
\verb{\{1\%, 2.5\%, 5\%, 10\%\}} significance levels
\item table \code{UDmax} reporting Double Max tests with critical values for
\verb{\{1\%, 2.5\%, 5\%, 10\%\}} significance levels.
}
}
\examples{
supF = dotest('inf','inflag',data=nkpc)
print(supF)

}
