% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Procedures.R
\name{print.seqtests}
\alias{print.seqtests}
\title{Print sequential SupF tests}
\usage{
\method{print}{seqtests}(x, ...)
}
\arguments{
\item{x}{\code{seqtests} class object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
No return value, only for printing formatted \code{seqtests} class object to console
}
\description{
\code{print} prints the object of class \code{seqtests} with the following information
\itemize{
\item Maximum number of breaks \code{m} in the tests
\item \code{sfl} table with sequential sup F tests statistics of l versus
l+1 breaks up to \code{m} breaks}
}
\examples{
seq_supF = doseqtests('inf','inflag',data=nkpc)
print(seq_supF)
}
