% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{pslr00}
\alias{pslr00}
\title{Test stat for  0 vs M breaks in mean given n=0 breaks in var}
\usage{
pslr00(y, M, trm, z, x = matrix(0, 0, 0), control = list())
}
\description{
Computes the UDmaxLRT test statistic for M coefficient changes given no variance changes
}
\details{
Note: This code is a translation of the one originally written by Pierre 
Perron and Yohei Yamamoto for MATLAB. Original code files can be found on 
Pierre Perron's website: https://blogs.bu.edu/perron/codes/
}
\references{
Perron, Pierre, Yohei Yamamoto, and Jing Zhou (2020), "Testing Jointly for Structural Changes in the Error Variance and Coefficients of a Linear Regression Model" \emph{Quantitative Economics}, vol 11, 1019-1057.
}
