% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{pslr10}
\alias{pslr10}
\title{Test stat for  n vs n+1 breaks in var given m breaks in mean}
\usage{
pslr10(y, m, n, h, z, x = matrix(0, 0, 0), control = list())
}
\description{
Computes the SeqLR_10T test statistic for n variance changes versus n + 1 variance changes given m coefficient changes
}
\details{
Note: This code is a translation of the one originally written by Pierre 
Perron and Yohei Yamamoto for MATLAB. Original code files can be found on 
Pierre Perron's website: https://blogs.bu.edu/perron/codes/
}
\references{
Perron, Pierre, Yohei Yamamoto, and Jing Zhou (2020), "Testing Jointly for Structural Changes in the Error Variance and Coefficients of a Linear Regression Model" \emph{Quantitative Economics}, vol 11, 1019-1057.
}
