% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{pslr2}
\alias{pslr2}
\title{Test stat for  0 vs n breaks in variance given M=m breaks in mean}
\usage{
pslr2(y, m, n, trm, z, x = matrix(0, 0, 0), control = list())
}
\value{
suplr - corrected test statistic (see eq. 9 of PYZ 2020)
}
\description{
Computes the sup LR_2T test statistic for n variance changes given m coefficient changes
}
\details{
Note: This code is a translation of the one originally written by Pierre 
Perron and Yohei Yamamoto for MATLAB. Original code files can be found on 
Pierre Perron's website: https://blogs.bu.edu/perron/codes/
}
\references{
Perron, Pierre, Yohei Yamamoto, and Jing Zhou (2020), "Testing Jointly for Structural Changes in the Error Variance and Coefficients of a Linear Regression Model" \emph{Quantitative Economics}, vol 11, 1019-1057.
}
