% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointtests.R
\name{pslr3}
\alias{pslr3}
\title{Test stat for  0 vs M breaks in mean given N=n breaks in var}
\usage{
pslr3(y, m, n, h, z, x = matrix(0, 0, 0), control = list())
}
\description{
Computes the sup LR_3T test statistic for m coefficient changes given n variance changes
}
\details{
Note: This code is a translation of the one originally written by Pierre 
Perron and Yohei Yamamoto for MATLAB. Original code files can be found on 
Pierre Perron's website: https://blogs.bu.edu/perron/codes/
}
\references{
Perron, Pierre, Yohei Yamamoto, and Jing Zhou (2020), "Testing Jointly for Structural Changes in the Error Variance and Coefficients of a Linear Regression Model" \emph{Quantitative Economics}, vol 11, 1019-1057.
}
