% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{pslr5}
\alias{pslr5}
\title{Test stat for  0 vs N breaks in variance given m=0 breaks in mean}
\usage{
pslr5(y, N, trm, z, x = matrix(0, 0, 0), control = list())
}
\description{
Computes the UDmaxLRT_1 test statistic for n variance changes given no coefficient changes
}
\details{
Note: This code is a translation of the one originally written by Pierre 
Perron and Yohei Yamamoto for MATLAB. Original code files can be found on 
Pierre Perron's website: https://blogs.bu.edu/perron/codes/
}
\references{
Perron, Pierre, Yohei Yamamoto, and Jing Zhou (2020), "Testing Jointly for Structural Changes in the Error Variance and Coefficients of a Linear Regression Model" \emph{Quantitative Economics}, vol 11, 1019-1057.
}
