% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{pslr9}
\alias{pslr9}
\title{Test stat for  m vs m+1 breaks in mean given n breaks in var}
\usage{
pslr9(y, m, n, h, z, x = matrix(0, 0, 0), control = list())
}
\description{
Computes the SeqLR_9T test statistic for m coefficient changes versus m + 1 coefficient changes given n variance change
}
\details{
Note: This code is a translation of the one originally written by Pierre 
Perron and Yohei Yamamoto for MATLAB. Original code files can be found on 
Pierre Perron's website: https://blogs.bu.edu/perron/codes/
}
\references{
Perron, Pierre, Yohei Yamamoto, and Jing Zhou (2020), "Testing Jointly for Structural Changes in the Error Variance and Coefficients of a Linear Regression Model" \emph{Quantitative Economics}, vol 11, 1019-1057.
}
