% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_methods.R
\name{pvdel}
\alias{pvdel}
\title{Covariance matrix of estimates delta.}
\usage{
pvdel(y, z, i, q, Tsize, b, prewhit, robust, x, p, withb, hetdat, hetvar)
}
\description{
procedure that compute the covariance matrix of the estimates delta.
}
\details{
Note: This code is an adaptation of the one originally written by Yohei 
Yamamoto and Pierre Perron for MATLAB. Original codes can be found on 
Pierre Perron's website: https://blogs.bu.edu/perron/codes/
}
\references{
Bai, Jushan & Pierre Perron (1998), "Estimating and Testing Linear Models with Multiple Structural Changes," \emph{Econometrica}, vol 66, 47-78.

Bai, Jushan & Pierre Perron (2003), "Computation and Analysis of Multiple Structural Change Models," \emph{Journal of Applied Econometrics}, 18, 1-22.
}
