% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{real}
\alias{real}
\title{World Health Organization TB data}
\format{
## `real`
A data frame with 103 rows and 1 column:
\describe{
  \item{rate}{Real exchange rate}
}
}
\source{
Garcia, R. and Perron, P., 1996. "An analysis of the real interest rate under regime shifts." Review of Economics and Statistics 111–125.
}
\usage{
real
}
\description{
Data set from the Garcia and Perron study's of ex-post exchange rate.
}
\keyword{datasets}
