% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rint}
\alias{rint}
\title{US Real interest rate 1961Q1 - 1986Q3}
\format{
\describe{US ex post real interest rate from 1961Q1–1986Q3 used in Perron, Yamamoto, and Zhou (2020), Garcia and Perron (1996), and Bai and Perron (2003)
  \item{rint}{US real interest rate series}
}
}
\source{
\url{https://blogs.bu.edu/perron/codes/}
}
\usage{
rint
}
\description{
US Real interest rate 1961Q1 - 1986Q3
}
\references{
Bai, J. and P. Perron (2003), "Computation and analysis of multiple structural change models." \emph{Journal of Applied Econometrics}, vol 18, 1–22.

Garcia, Rene and Pierre Perron (1996), "An analysis of the real interest rate under regime shifts." \emph{Review of Economics and Statistics}, vol 78, 111–125.

Perron, Pierre, Yohei Yamamoto, and Jing Zhou (2020), "Testing Jointly for Structural Changes in the Error Variance and Coefficients of a Linear Regression Model" \emph{Quantitative Economics}, vol 11, 1019-1057.
}
\keyword{datasets}
