% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structbreaks.R
\name{segmake}
\alias{segmake}
\title{Find segments}
\usage{
segmake(K, brk, m, n, cbrind, vbrind, q)
}
\description{
Find segments
}
\references{
Perron, Pierre, Yohei Yamamoto, and Jing Zhou (2020), "Testing Jointly for Structural Changes in the Error Variance and Coefficients of a Linear Regression Model" \emph{Quantitative Economics}, vol 11, 1019-1057.
}
