% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.R
\name{sequa}
\alias{sequa}
\title{sequential procedure to obtain number of breaks and break dates}
\usage{
sequa(m, signif, q, h, bigT, robust, prewhit, z, y, x, p, hetdat, hetvar, eps1)
}
\description{
sequential procedure to obtain number of breaks and break dates
}
\references{

}
