% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ssr_vec}
\alias{ssr_vec}
\title{Compute SSR recursively}
\usage{
ssr_vec(start, y, z, h, last)
}
\arguments{
\item{start}{starting entry of the sample used.}

\item{y}{dependent variable}

\item{z}{matrix of regressors of dimension q}

\item{h}{minimal length of a segment}

\item{last}{ending date of the last segment considered.}
}
\description{
This procedure computes recursive residuals from a data set that starts
at date "start" and ends at date "last". It returns a vector of sum of
squared residuals (SSR) of length last-start+1 (stored for convenience in a vector of length T).
}
\details{
Note: This code is an adaptation of the one originally written by Yohei 
Yamamoto and Pierre Perron for MATLAB. Original codes can be found on 
Pierre Perron's website: https://blogs.bu.edu/perron/codes/
}
\references{
Bai, Jushan & Pierre Perron (1998), "Estimating and Testing Linear Models with Multiple Structural Changes," \emph{Econometrica}, vol 66, 47-78.

Bai, Jushan & Pierre Perron (2003), "Computation and Analysis of Multiple Structural Change Models," \emph{Journal of Applied Econometrics}, 18, 1-22.
}
