% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_methods.R
\name{vec}
\alias{vec}
\title{Vectorize}
\usage{
vec(a)
}
\description{
procedure to jointly estimates the coefficient and variance break dates following section 5.1 of Qu and Perron (2007). The code follows est.m of Qu and Perron (2007)
}
\references{
Perron, Pierre, Yohei Yamamoto, and Jing Zhou (2020), "Testing Jointly for Structural Changes in the Error Variance and Coefficients of a Linear Regression Model" \emph{Quantitative Economics}, vol 11, 1019-1057.
}
