% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Procedures.R
\name{compile_seqtests}
\alias{compile_seqtests}
\title{Compile the output of sequential Sup Wald test}
\usage{
compile_seqtests(x)
}
\arguments{
\item{x}{`seqtests` class object}
}
\value{
class `seqtests` list `x` with appended data frame `sfl` containing the
sequential SupF test statistics with critical values at
\emph{10\%, 5\%, 2.5\%, and 1\%} level.
}
\description{
`compile_seqtests` formats the output of the `seqtests` class object to 1 table
\describe{
\item{`sfl`}{A table containing sequential sup F tests statistics
of `l` versus `l+1` for `l` in `1` to `m` breaks
with critical values of the corresponding tests at
\emph{1\%, 2.5\%, 5\%, and 10\%} significance levels.}
}
}
