% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstDate.R
\name{dating}
\alias{dating}
\title{Calculate optimal break points for linear dating procedure}
\usage{
dating(y, z, h, m, q, bigT)
}
\arguments{
\item{y}{dependent variable}

\item{z}{matrix of independent variables}

\item{h}{minimum length of segment}

\item{m}{maximum number of breaks}

\item{q}{number of regressors z}

\item{bigT}{sample period T}
}
\value{
A list containing the following components:
\item{glb}{minimum global SSR}
\item{datevec}{Vector of dates (optimal minimizers)}
\item{bigvec}{Associated SSRs}
}
\description{
Function computes break points that globally minimizes SSR
}
