% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Procedures.R
\name{dorepart}
\alias{dorepart}
\title{Repartition procedure}
\usage{
dorepart(
  y_name,
  z_name,
  x_name,
  data,
  m = 5,
  eps = 1e-05,
  eps1 = 0.15,
  maxi = 10,
  fixb = 0,
  betaini = 0,
  printd = 0,
  prewhit = 1,
  robust = 1,
  hetdat = 1,
  hetvar = 1
)
}
\arguments{
\item{m}{maximum number of breaks}

\item{eps}{convergence criterion for iterative recursive computation}

\item{eps1}{trimming level}

\item{maxi}{maximum number of iterations}

\item{fixb}{option to use fixed initial input \eqn{\beta}. If \code{1},
the model will use values given in \code{betaini}. If \code{0}, betaini is skipped}

\item{betaini}{Initial \eqn{beta_0} to use in estimation}

\item{printd}{option to print results of iterations for partial change model}

\item{prewhit}{option to use AR(1) for prewhitening process}

\item{robust, hetdat, hetvar}{options on error terms assumptions}

\item{y}{dependent variables in matrix form}

\item{z}{matrix of independent variables with coefficients are allowed to change across
regimes}

\item{x}{matrix of independent variables with coefficients constant across regimes}
}
\value{
reparv Repartition method estimation of break dates
}
\description{
The following procedure constructs the so-called repartition
estimates of the breaks obtained by the sequential method (see Bai
(1995), Estimating Breaks one at a time, Econometric Theory, 13,
315-352. It alows estimates that have the same asymptotic
distribution as those obtained by global minimization. Otherwise, the
output from the procedure "estim" below do not deliver asymptotically
correct confidence intervals for the break dates.
}
