% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Procedures.R
\name{doseqtests}
\alias{doseqtests}
\title{SupF(l+1|l) test}
\usage{
doseqtests(
  y_name,
  z_name = NULL,
  x_name = NULL,
  data,
  m = 5,
  eps = 1e-05,
  eps1 = 0.15,
  maxi = 10,
  fixb = 0,
  betaini = 0,
  printd = 0,
  prewhit = 1,
  robust = 1,
  hetdat = 1,
  hetvar = 1,
  const = 1
)
}
\arguments{
\item{m}{maximum number of breaks}

\item{eps}{convergence criterion for iterative recursive computation}

\item{eps1}{trimming level}

\item{maxi}{maximum number of iterations}

\item{fixb}{option to use fixed initial input \eqn{\beta}. If \code{1},
the model will use values given in \code{betaini}. If \code{0}, betaini is skipped}

\item{betaini}{Initial \eqn{beta_0} to use in estimation}

\item{printd}{option to print results of iterations for partial change model}

\item{prewhit}{option to use AR(1) for prewhitening}

\item{robust, hetdat, hetvar}{options on error terms assumptions}

\item{y}{dependent variables in matrix form}

\item{z}{matrix of independent variables with coefficients are allowed to change across
regimes}

\item{x}{matrix of independent variables with coefficients constant across regimes}
}
\value{
A list that contains following:
\itemize{
\item{supfl: SupF(l+1|l) test statistics}
\item{cv: Critical values for SupF(l+1|l) test}
\item{ndat: New date (if available)} }
}
\description{
Function computes the procedure of SupF(l+1|l) test. The function returns
the test statistics of supF(l+1|l) test
with null hypothesis is maximum number of break is l
and alternative hypothesis is l+1.
The l breaks under the null hypothesis are taken from the
global minimization. Also, new date (if available) and critical values based on
significant levels are returned for plotting and inference
}
