% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Procedures.R
\name{dotest}
\alias{dotest}
\title{SupF, UDMax & WDMax testing procedure}
\usage{
dotest(
  y_name,
  z_name = NULL,
  x_name = NULL,
  data,
  m = 5,
  eps = 1e-05,
  eps1 = 0.15,
  maxi = 10,
  fixb = 0,
  betaini = 0,
  printd = 0,
  prewhit = 1,
  robust = 1,
  hetdat = 1,
  hetvar = 1
)
}
\arguments{
\item{y_name}{name of dependent variable in the data set}

\item{z_name}{name of independent variables in the data set which coefficients are allowed to change
across regimes. \code{default} is vector of 1 (Mean-shift model)}

\item{x_name}{name of independent variables in the data set which coefficients are constant across
regimes. \code{default} is NULL}

\item{data}{the data set for estimation}

\item{m}{maximum number of breaks}

\item{eps}{convergence criterion for iterative recursive computation}

\item{eps1}{trimming level}

\item{maxi}{maximum number of iterations}

\item{fixb}{option to use fixed initial input \eqn{\beta}. If \code{1},
the model will use values given in \code{betaini}. If \code{0}, betaini is skipped}

\item{betaini}{Initial \eqn{beta_0} to use in estimation}

\item{printd}{option to print results of iterations for partial change model}

\item{prewhit}{option to use AR(1) for prewhitening}

\item{robust, hetdat, hetvar}{options on error terms assumptions}
}
\value{
A list that contains following:
\itemize{
\item{ftest: SupF test of 0 vs m (1 to maximum) breaks statistics}
\item{cv_supF: Critical values for Sup F test }
\item{cv_Dmax: Critical values for Double Max test}
\item{supF1: table summarizing the SupF test (for viewing purposes)}
\item{UDMax: table summarizing the Double Max test (including UDMax statistics and CVs)}
}
}
\description{
The procedure calculate the test statistics and print results of the 2 main tests:
\itemize{
\item{SupF test} {F test of 0 vs m breaks}
\item{Double Max test} {UDMax: the unweighted version
and WDMax: the weighted version}
}
}
