% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_methods.R
\name{estim}
\alias{estim}
\title{Estimate model with breaks}
\usage{
estim(y, z, x, m, b, robust, prewhit, hetomega, hetq, hetdat, hetvar)
}
\description{
This function estimates model after breaks are found. It also 
computes and reports confidence intervals for the break dates. The method used 
depends on the specification for robust
}
\details{
Note: This code is an adaptation of the one originally written by Yohei 
Yamamoto and Pierre Perron for MATLAB. Original codes can be found on 
Pierre Perron's website: https://blogs.bu.edu/perron/codes/
}
\references{
Bai, Jushan & Pierre Perron (1998), "Estimating and Testing Linear Models with Multiple Structural Changes," \emph{Econometrica}, vol 66, 47-78.

Bai, Jushan & Pierre Perron (2003), "Computation and Analysis of Multiple Structural Change Models," \emph{Journal of Applied Econometrics}, 18, 1-22.
}
