% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structbreaks.R
\name{getcv}
\alias{getcv}
\title{Get critical values using equations from Bai Perron (2003)}
\usage{
getcv(alpha, q, k, eps)
}
\description{
Get critical values using equations from Bai Perron (2003)
}
\references{
Bai, J. and Perron, P. (2003), Critical values for multiple structural change tests. \emph{The Econometrics Journal}, 6: 72-78. https://doi.org/10.1111/1368-423X.00102
}
