% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Variance.R
\name{interval}
\alias{interval}
\title{Break interval
Function computes confidence intervals for the break dates}
\usage{
interval(y, z, zbar, b, q, m, robust, prewhit, hetomega, hetq, x, p)
}
\arguments{
\item{y}{dependent variable}

\item{z}{matrix of independent variables with coefficients allowed to change
across regimes}

\item{zbar}{partitioned matrix of independent variables with coefficients allowed
to change across regimes according to break date}

\item{b}{vector of estimated dates of breaks}

\item{q}{number of regressors \code{z}}

\item{m}{maximum number of breaks}

\item{robust, withb, hetdat, hetvar}{options for assumptions of error terms
structure}

\item{prewhit}{Option of using prewhitening process. If \code{1}, an AR(1)
process will be used to filter. If \code{0}, skipped the filtering process}

\item{x}{matrix of independent variables with coefficients constant
across regimes}

\item{p}{number x regressors}

\item{bigT}{sample period T}
}
\value{
bound Confidence intervals of break date in \code{90\%} and \code{95\%}
significant level
}
\description{
Function to compute confidence intervals for the break dates based on
the "shrinking shifts" asymptotic framework
}
