\name{mbreaks-package}
\alias{mbreaks-package}
\alias{mbreaks}
\docType{package}
\title{
\packageTitle{mbreaks}
}
\description{
\packageDescription{mbreaks}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mbreaks}
\packageIndices{mbreaks}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{mbreaks}

Maintainer: \packageMaintainer{mbreaks}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
