% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl.R
\name{mdl}
\alias{mdl}
\title{Robust structural change estimation}
\usage{
mdl(
  y_name,
  z_name = NULL,
  x_name = NULL,
  data,
  eps1 = 0.15,
  m = -1,
  prewhit = 1,
  robust = 1,
  hetdat = 1,
  hetvar = 1,
  hetomega = 1,
  hetq = 1,
  maxi = 10,
  eps = 1e-05,
  fixn = -1,
  printd = 0
)
}
\arguments{
\item{y_name}{name of dependent variable in the data set}

\item{z_name}{name of independent variables in the data set which coefficients are allowed to change
across regimes. \code{default} is vector of 1 (Mean-shift model)}

\item{x_name}{name of independent variables in the data set which coefficients are constant across
regimes. \code{default} is NULL}

\item{data}{the data set for estimation}

\item{eps1}{value of trimming (in percentage) for the construction
and critical values. Minimal segment length will be set
at \code{default} = int(\code{eps1}*T) (T is total sample size).
\itemize{
\item{\code{eps1}=.05: Maximal value of \code{m} = 10}
\item{\code{eps1}=.10: Maximal value of \code{m} = 8}
\item{\code{eps1}=.15: Maximal value of \code{m} = 5}
\item{\code{eps1}=.20: Maximal value of \code{m} = 3}
\item{\code{eps1}=.25: Maximal value of \code{m} = 2}
} The \code{default} value is set at 0.15}

\item{m}{Maximum number of structural changes allowed. If not specify,
m will be set to \code{default} value from \code{eps1}}

\item{prewhit}{set to \code{1} if want to apply AR(1) prewhitening prior to estimating
the long run covariance matrix.}

\item{robust}{set to \code{1} if want to allow for heterogeneity
and autocorrelation in the residuals, \code{0} otherwise.
The method used is \emph{Andrews(1991)} automatic bandwidth with AR(1) approximation and the quadratic
kernel. Note: Do not set to \code{1} if lagged dependent variables are
included as regressors.}

\item{hetdat}{option for the construction of the F tests. Set to 1 if want to
allow different moment matrices of the regressors across segments.
If \code{hetdat} = \code{0}, the same moment matrices are assumed for each segment
and estimated from the ful sample. It is recommended to set
\code{hetdat}=\code{1} if number of regressors \code{x} > \code{0}.}

\item{hetvar}{option for the construction of the F tests.Set to \code{1}
if users want to allow for the variance of the residuals to be different across segments.
If \code{hetvar}=\code{0}, the variance of the residuals is assumed constant
across segments and constructed from the full sample. \code{hetvar}=\code{1} when \code{robust} =\code{1})}

\item{hetomega}{used in the construction of the confidence intervals for the break
dates. If \code{hetomega}=\code{0}, the long run covariance matrix of zu is
assumed identical across segments
(the variance of the errors u if \code{robust}={0})}

\item{hetq}{used in the construction of the confidence intervals for the break
dates. If \code{hetq}=\code{0}, the moment matrix of the data is assumed identical
across segments}

\item{maxi}{number of maximum iterations for recursive calculations of finding
global minimizers.\code{default} = 10}

\item{eps}{convergence criterion for recursive calculations}

\item{fixn}{number of pre-specified breaks. \code{default} = -1. It will be replaced
automatically to 2 if no specification is given}

\item{printd}{Print option for model estimation. \code{default} = 0, to
suppress intermediate outputs printing to console}
}
\value{
A list that contains the following:
\itemize{
\item{Wtest: Sup F tests of 0 versus m breaks and Double Max tests}
\item{spflp1: Sequential Sup F test of l versus l+1 breaks }
\item{BIC: Estimated number of breaks by BIC and the corresponding model}
\item{LWZ: Estimated number of breaks by LWZ and the corresponding model}
\item{sequa: Estimated number of breaks by sequential procedure the corresponding model}
\item{repart: Estimated number of breaks by repartition procedure the corresponding model}
\item{fix: Estimated model with pre-specified number of breaks}
}
}
\description{
Function executes all the main procedures to estimate either i) pure structural
breaks model or ii) partial structural breaks model
}
\details{
The 7 main procedures include:
\itemize{
\item{\code{\link{dotest}}: Procedure to conduct SupF test of 0 versus m breaks and
Double Max test.}
\item{\code{\link{dospflp1}}: Procedure to conduct sequential SupF(l+1|l) (l versus l+1 breaks)}
\item{\code{\link{doorder}}: Procedure to find number of break by criteria selection
(including BIC and LWZ criterion)}
\item{\code{\link{dosequa}}: Procedure to find number of break and break dates via sequential method}
\item{\code{\link{dorepart}}: Procedure to find number of break and break dates via repartition method}
\item{\code{fix}: Procedure to find number of break by pre-specified number of breaks, which are set at \code{default}
to be 2 }}

All \code{default} values of error assumptions (\code{robust},
\code{hetdat}, \code{hetvar}, \code{hetq}) will be set to 1. The implications on
the structure of model's errors related to individual settings are explained within
the arguments section for each option. Users can separately invoke only one at a
time one of the main7 procedures mentioned above
}
