% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nkpc}
\alias{nkpc}
\title{New Keynesian Phillips curve data}
\format{
## `nkpc`
A data frame with 151 rows and 12 columns:
\describe{
  \item{year}{Current period year}
  \item{quarter}{Quarter in current period year}
  \item{inf}{Inflation rate}
  \item{inflag}{Inflation rate in previous period}
  \item{inffut}{Expected inflation rate, taken as value of inflation rate of next period}
  \item{ygap}{Productivity output gap}
  \item{lbs}{}
  \item{lbslag}{}
  \item{spreadlag}{}
  \item{dwlag}{}
  \item{dcplag}{}
}
}
\source{
Perron, P. and Yamamoto, Y., 2015. "Using ols to estimate and test for structural changes in models with endogenous regressors." Journal of Applied Econometrics 30, 119–144.
}
\usage{
nkpc
}
\description{
Data set from inflation and other macroeconomic variables
}
\keyword{datasets}
