% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstDate.R
\name{nldat}
\alias{nldat}
\title{Non-linear dating procedure}
\usage{
nldat(y, z, x, h, m, p, q, bigT, fixb, eps, maxi, betaini, printd)
}
\arguments{
\item{y}{time-series observations of dependent variable}

\item{z}{matrix of regressors which coefficients are allowed to change across regimes}

\item{x}{matrix of regressors which coefficients are constant across regime}

\item{p}{number of \code{z} regressors}

\item{q}{number of \code{x} regressors}

\item{bigT}{the sample size T}

\item{fixb}{option to use initial \eqn{\beta} If \code{1}, procedure requires \code{betaini}.
If \code{0}, procedure will not use initial beta values}

\item{eps}{Convergence criterion}

\item{betaini}{initial beta values. Required when use with option \code{fixb}}

\item{printd}{option to print output from iterated estimations. If \code{1}, the results
for each iteration will be printed in console log. If \code{0}, no output will be printed}
}
\value{
A list containing the following components:
\item{glb}{minimum global SSR}
\item{datevec}{Vector of dates (optimal minimizers)}
\item{bigvec}{Associated SSRs}
}
\description{
Function computes the break dates of a partial structural change model
with pre-specified number of breaks m. The procedure is iterations of
recursive computations of SSR and finding the m optimal global minimizers
}
\references{

}
