% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstDate.R
\name{parti}
\alias{parti}
\title{Calculate optimal point for segment partition}
\usage{
parti(start, b_start, b_end, last, bigvec, bigT)
}
\arguments{
\item{start}{start date index of the segment}

\item{b_start}{first possible break date}

\item{b_end}{last possible breakdate}

\item{last}{end date index of the segment}

\item{bigT}{sample period T}
}
\value{
A list containing the following components:
\item{ssrmin}{associated SSR of optimal break}
\item{dx}{optimal date (global minimizer)}
}
\description{
Function calculates the optimal one break partition for a segment starting
from start to last. The possible range of the break is within
\eqn{[b_{last},b_{end}]} based on the vector of recursive SSR of the model
}
\references{

}
