% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Procedures.R
\name{plot_model}
\alias{plot_model}
\title{#' Plot model of estimated n breaks
plot.model = function(x,...){
#get data from the model
m = x$nbreak
y = x$y
zreg = x$z
xreg = x$x
p = x$numx
q = x$numz
date = x$date
beta = x$beta
zbar = diag_par(zreg,m,date)
T = length(y)
if (p == 0){
reg = zbar
}
else{
reg = cbind(xreg,zbar)
}

#compute model with no breaks
fixreg = cbind(xreg,zreg)
fixbeta = OLS(y,fixreg)
fity_fix = fixreg\%*\%fixbeta

fity = reg\%*\%beta
tx = seq(1,T,1)

range_y = max(y)-min(y);

#plot original series
graphics::plot(tx,y,type='l',col="black", xlab='time',ylab="y",
ylim=c(min(y)-range_y/10,max(y)+range_y/10),lty=1)

#plot fitted values series

graphics::lines(tx, fity,type='l', col="blue",lty=2)

#plot fitted values series

graphics::lines(tx, fity_fix,type='l', col="dark red",lty=2)

for (i in 1:m){
graphics::abline(v=date\link{i,1},lty=2)
if (x$CI\link{i,1} < 0 || x$CI\link{i,2}>T){}else{
graphics::segments(x$CI\link{i,1},min(y)\emph{(12+i/m)/10,x$CI\link{i,2},min(y)}(12+i/m)/10,lty=1,col='red')}
}

legend(0,max(y)+range_y/10,legend=c("observed y",paste(m,'break y'),"0 break y"),
lty=c(1,2,2), col=c("black","blue","red"), ncol=1)

}}
\usage{
plot_model(model, CI = 0.95, title = NULL)
}
\description{
new plot function for class model
}
