% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl.R
\name{print.mdl}
\alias{print.mdl}
\title{Print information of `mbreaks` comprehensive procedure}
\usage{
\method{print}{mdl}(x, ...)
}
\arguments{
\item{x}{class `mdl` object}

\item{...}{further arguments passed to or from other methods}
}
\value{
No return value, only for printing `model`, `sbtests` and `seqtests` class objects
invoked during `mdl()`.
}
\description{
`print` prints the class `mdl` object with default showing only
certain procedures called
by `mdl()` function including: `seqtests` class object, `sbtests` class
object, and `model` class object using KT information criterion
}
\examples{
rate = mdl('rate',data=real)
print(rate)
}
