% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Procedures.R
\name{print.model}
\alias{print.model}
\title{Summary output of a n breaks model}
\usage{
\method{print}{model}(x, digits = -1, ...)
}
\arguments{
\item{x}{independent variables with regime-wise coefficients}

\item{nbreak}{number of breaks in the model}

\item{glb}{global minimum SSR of the model}

\item{datevec}{estimated break date}

\item{y}{dependent variables}

\item{z}{independent variables with regime-specific coefficients}
}
\value{
tbl Tables containings:
i) global minimum SSR
ii) estimated date
iii) estimated coefficients
}
\description{
Function to format the output of the n-break model
}
