% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Procedures.R
\name{print.sbtests}
\alias{print.sbtests}
\title{S3 print function for sup F tests
#'Function to format the output of the Sup F test}
\usage{
\method{print}{sbtests}(x, ...)
}
\description{
S3 print function for sup F tests
#'Function to format the output of the Sup F test
}
