% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Variance.R
\name{pvdel}
\alias{pvdel}
\title{Covariance matrix of estimator delta construction}
\usage{
pvdel(y, z, i, q, bigT, b, prewhit, robust, x, p, withb, hetdat, hetvar)
}
\arguments{
\item{y}{dependent variable}

\item{z}{matrix of independent variables with coefficients allowed to change
across regimes}

\item{i}{maximum number of breaks}

\item{q}{number of regressors \code{z}}

\item{bigT}{sample period T}

\item{b}{vector of estimated dates of breaks}

\item{prewhit}{Option of using prewhitening process. If \code{1}, an AR(1)
process will be used to filter. If \code{0}, skipped the filtering process}

\item{robust, withb, hetdat, hetvar}{options for assumptions of error terms
structure}

\item{x}{atrix of independent variables with constant coefficients
across regimes}

\item{p}{number x regressors}
}
\value{
vdel Covariance matrix of delta
}
\description{
Covariance matrix of estimator delta construction
}
